rename nc_vector.h to vector2.h
