+++ nc_runtime.vcxproj
@@ -88,7 +88,7 @@
     <ClInclude Include="include\nc_runtime\nc_stdlib.h" />
     <ClInclude Include="include\nc_runtime\nc_string.h" />
     <ClInclude Include="include\nc_runtime\nc_types.h" />
-    <ClInclude Include="include\nc_runtime\nc_vector.h" />
+    <ClInclude Include="include\nc_runtime\vector2.h" />
     <ClInclude Include="include\nc_runtime\string_slice.h" />
     <ClCompile Include="src\costmap_inflator_imple.cpp" />
     <ClCompile Include="src\nc_array.cpp" />
